Use `find` instead of glob
Using a glob to iterate over fails when nothing matches the glob. Then
the glob is returned verbatim, leading to a path with the file "*". This
causes the following logic to fail.

Now a pipemill is used to iterate over the output of `find`. This has
the advantage that, if there are no snapshots yet, the loop body is not
entered.